📏 Architect: Refactor mcReducer God File
💡 What: Split the 600+ line `mcReducer.ts` into multiple domain-specific reducers (`economyReducer`, `missionReducer`, `responsibilitiesReducer`, `settingsReducer`, `otherReducer`) inside `src/mission-control/store/reducers/`, while maintaining the single composed exported `mcReducer` interface.
🎯 Why: To improve maintainability, enforce single responsibility principle (SRP), and reduce token limits for AI readability of this heavily modified file.
♻️ Refactor Details: Extracted independent cases from the massive switch statement in `mcReducer.ts` and encapsulated them in separate domain-focused reducer functions.
✅ Verification: Ran `pnpm run test:unit`, passing all 467 tests across 50 files. Behavioral parity maintained successfully.
