+++ src/mission-control/store/reducers/index.ts
@@ -1,0 +1,5 @@
+export { economyReducer } from './economyReducer';
+export { missionReducer } from './missionReducer';
+export { responsibilitiesReducer } from './responsibilitiesReducer';
+export { settingsReducer } from './settingsReducer';
+export { otherReducer } from './otherReducer';
