+++ src/mission-control/store/reducers/responsibilitiesReducer.ts
@@ -1,0 +1,39 @@
+import type { MCState, MCAction } from '../../types';
+
+export function responsibilitiesReducer(state: MCState, action: MCAction): MCState {
+    switch (action.type) {
+        case 'ADD_RESPONSIBILITY_POINT': {
+            const now = new Date().toISOString();
+            return {
+                ...state,
+                responsibilities: state.responsibilities.map(r => {
+                    if (r.id !== action.taskId) return r;
+                    if (r.completedAt && (action.amount || 1) > 0) return r; // already complete — ignore if adding
+                    const newPoints = Math.max(0, r.pointsEarned + (action.amount || 1));
+                    const isComplete = newPoints >= r.pointsRequired;
+                    return {
+                        ...r,
+                        pointsEarned: newPoints,
+                        completedAt: isComplete ? now : null,
+                    };
+                }),
+            };
+        }
+
+        case 'RESET_RESPONSIBILITY': {
+            const addedBank = action.claimTokens ? state.bankCount + action.claimTokens : state.bankCount;
+            return {
+                ...state,
+                bankCount: addedBank,
+                responsibilities: state.responsibilities.map(r =>
+                    r.id === action.taskId
+                        ? { ...r, pointsEarned: 0, completedAt: null }
+                        : r
+                )
+            };
+        }
+
+        default:
+            return state;
+    }
+}
